made changes so that build works in linux #3

+++ src/MeshData.cpp
@@ -5,7 +5,7 @@ MeshData::MeshData(int l, int d)
 	data = new gameBlock[l * d];
 	vLength = l;
 	vDepth = d;
-	for (size_t i = 0; i < l * d; i++)
+	for (int i = 0; i < l * d; i++)
 	{
 		data[i] = gameBlock::UNSET;
 	}
